<refacoring> common/ & cobfigs/ name change
