+++ back/main_sever/src/_common/exception/common/enum/interceptor-exception.ts
