+++ back/main_sever/src/_common/exception/common/enum/uncatched-exceotion-code.enum.ts
