+++ back/main_sever/src/_common/exception/uncatch.exception.ts
