+++ back/main_sever/src/boards/entities/board.entity.ts
@@ -1,6 +1,6 @@
 import { IsIn, IsString } from 'class-validator';
-import { BaseModel } from 'src/common/entity/base.entity';
-import { stringValidationMessage } from 'src/common/validation-message/string-validation.message';
+import { BaseModel } from 'src/_common/entity/base.entity';
+import { stringValidationMessage } from 'src/_common/validation-message/string-validation.message';
 import { UserModel } from 'src/users/entities/user.entity';
 import { Column, Entity, JoinColumn, ManyToOne } from 'typeorm';
