+++ back/main_sever/src/posts/-comment/dto/create-comment.dto.ts
@@ -3,6 +3,6 @@ import { CommentModel } from '../entities/comment.entity';
 import { Column } from 'typeorm';
 import { IsBoolean, IsNotEmpty, IsString } from 'class-validator';
 import { Transform } from 'class-transformer';
-import { stringValidationMessage } from 'src/common/validation-message/string-validation.message';
+import { stringValidationMessage } from 'src/_common/validation-message/string-validation.message';
 
 export class CreateCommentDto extends PickType(CommentModel, ['content']) {}
