+++ back/main_sever/src/posts/-comment/dto/update-comment.dto.ts
@@ -1,6 +1,6 @@
 import { PartialType, PickType } from '@nestjs/mapped-types';
 import { IsBoolean, IsOptional, IsString } from 'class-validator';
-import { stringValidationMessage } from 'src/common/validation-message/string-validation.message';
+import { stringValidationMessage } from 'src/_common/validation-message/string-validation.message';
 import { CreateCommentDto } from './create-comment.dto';
 import { Column } from 'typeorm/decorator/columns/Column';
