+++ back/main_sever/src/posts/-comment/entities/comment.entity.ts
@@ -2,8 +2,8 @@ import { IsNotEmpty, IsNumber, IsString } from 'class-validator';
 import { ManyToOne } from 'typeorm/decorator/relations/ManyToOne';
 import { Column, Entity, JoinColumn } from 'typeorm';
 
-import { BaseModel } from 'src/common/entity/base.entity';
-import { stringValidationMessage } from 'src/common/validation-message/string-validation.message';
+import { BaseModel } from 'src/_common/entity/base.entity';
+import { stringValidationMessage } from 'src/_common/validation-message/string-validation.message';
 import { UserModel } from 'src/users/entities/user.entity';
 import { QuestPostModel } from 'src/posts/entities/quest_post.entity';
 import { SbPostModel } from 'src/posts/entities/sb_post.entity';
