+++ back/main_sever/src/posts/-quest/quest.controller.ts
@@ -15,13 +15,13 @@ import { PostsUseCases } from '../usecase/post.use-case';
 import { BoardUseCase } from 'src/boards/usecase/board.use-case';
 import { User } from 'src/users/decorator/user.decorator';
 import { CreatePostDto } from '../dto/create-post.dto';
-import { QueryRunner } from 'src/common/decorator/query-runner.decorator';
+import { QueryRunner } from 'src/_common/decorator/query-runner.decorator';
 import { QueryRunner as QR } from 'typeorm';
 import { AccessTokenGuard } from 'src/auth/guard/token/access-token.guard';
-import { TransactionInterceptor } from 'src/common/interceptor/transaction.interceptor';
+import { TransactionInterceptor } from 'src/_common/interceptor/transaction.interceptor';
 import { PostApiResponseDto } from 'hoodone-shared/dist/response-dto/post-api-reponse.dto';
 import { GetPaginatedPostsQueryDTO } from '../dto/get-paginated-posts.dto';
-import { IsPublic } from 'src/common/decorator/is-public.decorator';
+import { IsPublic } from 'src/_common/decorator/is-public.decorator';
 import { ParseIntPipe, ValidationPipe } from '@nestjs/common/pipes';
 import { Roles } from 'src/users/decorator/roles.decorator';
 import { RoleType } from 'src/users/const/role.type';
