+++ back/main_sever/src/posts/-sb-post/sb_post.controller.ts
@@ -14,11 +14,11 @@ import {
 import { QueryRunner as QR } from 'typeorm';
 
 import { AccessTokenGuard } from 'src/auth/guard/token/access-token.guard';
-import { TransactionInterceptor } from 'src/common/interceptor/transaction.interceptor';
-import { QueryRunner } from 'src/common/decorator/query-runner.decorator';
+import { TransactionInterceptor } from 'src/_common/interceptor/transaction.interceptor';
+import { QueryRunner } from 'src/_common/decorator/query-runner.decorator';
 import { User } from 'src/users/decorator/user.decorator';
 
-import { IsPublic } from 'src/common/decorator/is-public.decorator';
+import { IsPublic } from 'src/_common/decorator/is-public.decorator';
 import { Roles } from 'src/users/decorator/roles.decorator';
 import { RoleType } from 'src/users/const/role.type';
 import { Logger } from '@nestjs/common';
