+++ back/main_sever/src/posts/dto/update-post.dto.ts
@@ -2,7 +2,7 @@ import { PartialType, PickType } from '@nestjs/mapped-types';
 import { PostModel } from '../entities/post.entity';
 import { CreatePostDto } from './create-post.dto';
 import { IsOptional, IsString } from 'class-validator';
-import { stringValidationMessage } from 'src/common/validation-message/string-validation.message';
+import { stringValidationMessage } from 'src/_common/validation-message/string-validation.message';
 
 export class UpdatePostDto extends PartialType(CreatePostDto) {
     @IsString({
