+++ back/main_sever/src/users/entities/temp-user.entity.ts
@@ -1,10 +1,10 @@
 import { IsEmail, IsEnum, IsString, Length, Matches } from 'class-validator';
 import { Column, Entity, OneToMany } from 'typeorm';
 
-import { emailValidationMessage } from 'src/common/validation-message/email-validation.message';
-import { lengthValidationMessage } from 'src/common/validation-message/length-validation.message';
-import { stringValidationMessage } from 'src/common/validation-message/string-validation.message';
-import { BaseModel } from 'src/common/entity/base.entity';
+import { emailValidationMessage } from 'src/_common/validation-message/email-validation.message';
+import { lengthValidationMessage } from 'src/_common/validation-message/length-validation.message';
+import { stringValidationMessage } from 'src/_common/validation-message/string-validation.message';
+import { BaseModel } from 'src/_common/entity/base.entity';
 import { RoleType } from '../const/role.type';
 
 @Entity('temporaryUser')
