+++ back/main_sever/src/users/entities/user.entity.ts
@@ -3,11 +3,11 @@ import { Exclude, Expose } from 'class-transformer';
 import { IsEmail, IsEnum, IsString, Length, Matches } from 'class-validator';
 import { Column, Entity, OneToMany } from 'typeorm';
 
-import { emailValidationMessage } from 'src/common/validation-message/email-validation.message';
-import { lengthValidationMessage } from 'src/common/validation-message/length-validation.message';
-import { passwordValidationMessage } from 'src/common/validation-message/password-vaildation.message';
-import { stringValidationMessage } from 'src/common/validation-message/string-validation.message';
-import { BaseModel } from 'src/common/entity/base.entity';
+import { emailValidationMessage } from 'src/_common/validation-message/email-validation.message';
+import { lengthValidationMessage } from 'src/_common/validation-message/length-validation.message';
+import { passwordValidationMessage } from 'src/_common/validation-message/password-vaildation.message';
+import { stringValidationMessage } from 'src/_common/validation-message/string-validation.message';
+import { BaseModel } from 'src/_common/entity/base.entity';
 import { RoleType } from '../const/role.type';
 import { BoardModel } from 'src/boards/entities/board.entity';
 import { PostModel } from 'src/posts/entities/post.entity';
